Move config to its own package and respect redis config in integration tests
The main package is only affected by this change, since it now imports
config separately.
